Improve code logic & therefore better error messages
This commit deals with issue #31, doesnt really fix anything but makes
room for potential bugs in the future.

What could have broken is if there was a file before 00-index.json it
would have been loaded before and self = None would have broken things.

Now it straight up loads 00-index.json first and if it cant do that it
fails